move BoneVisualizer to test
